requirements.txt
discord.py[voice]>=2.7.1
python-dotenv>=1.0.0
yt-dlp>=2026.01.01